refactor(daemon): extract root kill identity helpers
